Move Redux reducers into store directory
